Make SubnetMismatchForPort extend BadRequest rather than Conflict
Nova is trying to do a better job of mapping Neutron exceptions and
change I4c087684ef77988e5f463d7f2f50fc2a04f37db0 is trying to map 409 to
the proper nova exception. In looking at what raises a 409 exception in
Neutron, the SubnetMismatchForPort exception seemed like a candidate for
changing from Conflict (409) to BadRequest (400) to ease that mapping.

Note that the only thing currently using this exception is the L3 router
extension when removing a router interface and the 400 response code is
already listed in the API docs as a possible response code:

http://docs.openstack.org/api/openstack-network/2.0/content/router_remove_interface.html

Also note that it's generally OK to change APIs for a more accurate
response code:

https://wiki.openstack.org/wiki/APIChangeGuidelines#Generally_Considered_OK

Related-Bug: #1209446

Change-Id: I4507e1db69d738ec0f943f1b8b1209f269d5aebf