Fix IVS list ports method.
Fixes the IVS list ports method. There were two issues. First,
it was sending to stderr instead of stdout like the utils.execute
method expects by default. Second, it was calling 'list-ports',
which did not exist on IVS.

A new list-ports call is being added to IVS, but this adds fallback
code to handle parsing 'show' output as well.

Change-Id: Iae5c60b35c96e8a8341aff668b258889955bd27e