Disable connection keep-alive for thread-safety
The server manager object is instantiated when the plugin is first
initialized. When multiple workers are enabled, they all reference
the same servermanager object, so the connection ends up being shared
between threads, leading to mangling of the http connection.

This patch disables the keep-alive capability and prevents the connection
from being attached to the server manager to keep it thread-safe.

Change-Id: I0c7baf37fc4825ebdf4ba3deb00933a318367d30