Migrate to oslo logging from common log code
Neutron is switching to the new oslo logging library
in https://review.openstack.org/#/c/159638/. Once that is
merged, the neutron.openstack.common.log module will be
gone so this code needs to switch as well.

This adds oslo logging as a requirement since the requirements
haven't merged into master yet on Neutron. It can be removed
after Neutron has been updated.

Change-Id: Ie6b1d6973422dfa158fa10c540f084a0f1f92acc