Don't override status if it is already set
Floating IP create now provides a port status of 'N/A'
for the port it creates to allocate an IP to indicate
that the port isn't a real port. The monolithic plugin
needs to be updated to use that status if it's passed
in so it doesn't fail the unit test.

Change-Id: I796bd3adfe21de85eaf940948fe5d063ccc86778