Fix for DBDeadLock exceptions
The type of update query that uses multiple 'where' criteria
can result in internal MySQL deadlocks if there are a lot of
operations from other connections on the same table. These
are considered normal and the expectation is just that the
caller should re-run the query to try again.

The current code was just letting the exception go uncaught
though so instead of the update being retried, it would bubble
up and fail whatever request was going on.

Change-Id: I8906e1c1edf1478dcc6535dc8d053fcf7a0f4c86