Add support for NFV switch for BSN's ML2 mechanism driver
This patch adds support for NFV Switch type of virtual switch for the BSN
ML2 mechanism driver.

In the case of NFV Switch, when a bind_port() is called, the driver asks
the BCF controller to reserve a vhost socket (if available) for the
NFV endpoint. The reserved vhost socket info is used to populate the
vif_details, which shall in-turn be used to generate the required xml
for qemu to start the NFV endpoint.

Changes in the patch:
- bind_port() call determined by virtual switch type (as determined by BCF)
- implement bind_port for nfvswitch type compute nodes
- minor code cleanups

Corresponding NSAPI PR: https://github.com/bigswitch/bvs/pull/5070

Change-Id: I3055b2118e709e0ed4d73eb203988956b479ec9c