Several fixes to unblock gate
---------------------------------------------
Add missing requirement from test-requirements

One of the test modules imported from neutron depends
on testresources, which was not explicitly in test-requirements.txt.
This patch adds it. Unfortunately the dependency on neutron in the
requirements does not include the test-requirements.txt file.
---------------------------------------------
Don't inherit OVSBridge

We don't use anything from that class so there is no need to
inherit it.
---------------------------------------------
Skip L3 router interface dup test

The upstream logic wasn't anticipating being called in a transaction
due to it's rollback handling. This will be removed in
I5f473fff4f8372852d563c79dac2991089eb0b77 but the test can be skipped
for now.

Change-Id: I0cdd30ddac1c72b7302b56e52321bc0b315f6ca9