BVS-6743: Update BSN Agent for NFV type compute nodes
This patch adds support nfvswitch type of virtual switch for the bsn-agent.
This switch doesn't support list-ports as the virtual ports are statically
allocated during service start, hence no need to monitor it, resulting in the
functions to be no-op.

Also fixes a bug where-in the restproxy.ini file isn't being refered to
by the neutron-bsn-agent service

Change-Id: I319ffc585ad9e6e92ad08709f3c93c1e5be88561