BVS-6548: optimize the topo sync validation logic w.r.t. unsupported chars
Made some code clean-ups and optimizations for the name validation logic used
when generating the TopoSync blob. Instead of applying processing logic on the
dictionary, use it during the generation of the blob itself. This way, we not only
avoid and extra iterating over the map, but also skip a lot of extra processing
required to generate the map items (that are going to be excluded)

Change-Id: Icfbec66fd1106dae0cf06d015f219137d87e6a89
(cherry picked from commit 6f7e3cd1810ddc7c05017d6f631b08c63384e0b9)