BVS-7525: don't sync dangling objects to BCF
When a floating-ip is associated with a tenant, the tenant_id is that of the
associated tenant (not source tenant). When tenant is deleted, the floating-ips aren't.
Add a check to skip syncing floating-ips associated to non-existent tenants

Change-Id: I86acedd1da57f3bb174b1496d87834296f097fb5
(cherry picked from commit cc179e6a90bcf34b40685795b1e3c7c30d3724e4)