use context_manager for more transactions
There are a few places that are missing the new way of creating
transactions, and this was causing a number of UT to fail.

NOTE: This patch does not complete all of the work that is required
for the enginefacade switch. This patch is just enough to get the
failing UT to pass.

Change-Id: I52616f43c3b08d910784bbf003b22f0b2ab693d7