Add address scope info to network data
This commit adds support for the address scope feature in the BSN
plugin. This include UT.

The code changes result in extra data being added to the json blob
of data that is sent to the controller in the case of updating
or creating a network. The data that is added is:

 * A map of all of the address scopes and the subnetpools that are
   a part of those address scopes. This allows the controller to
   make any config changes that are required to support
 * a boolean flag "overlapping" which indicates whether or not there
   are any overlapping subnets in the address scopes. This is used
   by the controller in validation. Some versions of the controller
   will support overlapping address scopes, and some will not.

Change-Id: Ibb8b977580ac2b042c28bbac85392305e830abce