ubuntu 16.04 container doesn't need sudo
 - build container breaks with sudo
 - skips N344,N524,N530,N534 tox tests

Also does
Migrate to stestr and sphinx PTI commands

The PTI lists stestr as the test runner we're using now. Go ahead and
switch to it, as it simplifies many things.

Also, the PTI updated the sphinx commands, so update the docs env to
match what the gate will be doing.

Change-Id: I47aa85500a3ab92f970802914f17dc57e3b07f06