OSP-224: Cleanup for NSAPI context header
Intially the context header was intended for 2 purposes:

1. Update tenant name
2. Provide debug information to BCF

However, long ago, we changed to use tenant_name for identification and stops supporting renaming, so 1 is unessesary.

Note: we will support renaming with the next uuid change, but it will be using different approach and this context can not only be us
    ed for the purpose, but also causes intended behaviors to block it.

For 2, there might be some miscommunications among different devs as we have never been logging the information on BCF side.

Considering no one found this before, means that nobody uses this for debug, it is a waste of resource to generate the context for every REST call.

For above two reasons, we are removing related logics completely.

Change-Id: I8b536526b7515ac5587e3dcd52afaa660e462348