opt in for neutron-lib consumption patches
As part of the Denver PTG [1] we decided networking related projects that are "current" and want to receive neutron-lib consumption patches on an on-going basis should indicate such with a well defined comment in their requirements.txt. This allows us to easily find the list of project to receive neutron-lib consumption patches [2] by searching for the string. In addition, projects opting-in for these patches are also attesting they will stay up to date with TC and infra initiatives to ensure consumption patches can flow freely. This patch adds the "neutron-lib-current" string to requirements.txt opting in for neutron-lib consumption patches.

Change-Id: I1a430d3ac138a441164df6bb2391dad0a730a385