OSP-216 Add unicode display-name support, fix capability check, add/fix related unit tests
This commit adds an option to the plugin to enable/disable unicode based objects and store them on BCF.

When the config is enabled, the os_object id is used for bcf name, and os_object name is used for bcf display-name.

Default is unicode enabled (naming_scheme_unicode=True).

Unicode is enabled only when:
1. BCF supports it (5.0.0 or above)
2. naming_scheme_unicode is set to True in config (or if it is empty)

In other situations, unicode will be disabled and sent in old format.

---
As of Queens:

Objects that always have names:
- tenant(project)
- security-group

Objects that might not have names(Empty Names for these are now supported when unicode is enabled):
- endpoint(port)
- nat-profile(router)
- segment(network)

We don't care about names of other objects like floating ip or subnets.

---
This PR also does some change to capability check functions:

Before:
  - [] (empty capabilities) = fine
  - result capabilities = intersaction of two servers' capabilities
  - cached only once during startup

After:
  - [] (empty capabilities) = failed request, always query BCF to try update it
  - result capabilities = union of two servers' capabilities
  - update capabilities every 5 minutes (does not clear existing cache if it fails)

Note: even though capabilities is checked every 5 minutes, it only logs unicode enabled/disabled during service startup and when it changes. This is to reduce uneccesary logs.

---
This PR does not touch test_path/reachability test, which means it would only work when display-name is disabled currently.

There will be a later PR for it.

Change-Id: I9ecb13df063e85038b2a622724c874dec01b4bbc