OSP-244 Fix mock_response in TestBigSwitchMechDriverNetworksV2.test_create_network
Previously in test_create_network, there was a line of codes trying to modify http response status to 200, but the change was done after the http request was made. The code for modifiying the status also does not work. This change fixes this issue.

Also adds a print statement help with debug as the Error with test_create_network was not stable.

Change-Id: I7ad2c8777376bb6f82c7d313900b084424d5f0d3