OSP-244 Fix inheritance issue for TestBigSwitchMechDriverNetworksV2
TestBigSwitchMechDriverNetworksV2 was inherited from two classes,
both of them have a setUp() method, previously, TestNetworksV2's
setUp() was called, while our setUp() was not called. This caused
an issue that topo sync was still triggled and made
test_create_network unstable.

Changing the order of classes being inherited is easy but that would
not be compatible with python3.

Specifying the super class for setUp() is a more consistent fix.

Change-Id: Ic01394b4875d1cbafd84d26f093ad56ccfd424ab