OSP-244 Add Topo Sync Mock
We have made quite a few changes to the way we trigger topo sync,
and it can be triggered outside of watchdog currently.

Disabling the watchdog does not totally prevent topo sync from
being triggered.

It is easier to just return a success result than modifying codes
related to different cases that trigger topo sync, when there is
no need to verify those behaviors.

Currently the patch is only used in one test class:
TestBigSwitchMechDriverNetworksV2, but it can be reused in other
test classes if we find it neccessary.

Change-Id: If9f795507402cb870d93656b0f5b7f4fcb95c9d1