add local tox targets for pep8 and py27
This patch adds local tox targets for pep8 and python 2.7 as per [1]
that will install dependencies in editable mode.
To run them use the pep8-dev and py27-dev targets respectively.

Py3 support is still needed, but it appears there's lots of work
to do in this code before it'll run successfully.

[1] https://etherpad.openstack.org/p/neutron-sibling-setup

Change-Id: If65fb3494ca7453005be1526a5178ff2f581938e