OSP-255: Check status of test__notify_gateway_port_ip_changed
There was some error with the specific test.
We are not sure if it is a legit error or something needed to be
fixed uptream. We are skipping it to not block gating.

This PR tried to undo the changes that skips the test, which is
for verifying the status of the test.

Rebase against latest master to check if it still fails.

If it no longer fails, we can potentially merge it and resolve the
jira.

Change-Id: I2778a56dc1cb2e11d754d402167d2997447c2061