OSP-257: cleanup redundant packages from queens and onwards
we had to split the networking-bigswitch package into multiple
packages owing to the containerization of services in queens.

we now install the packages based on requirement on a per
container basis - instead of the previous approach of installing
everything as part of networking-bigswitch.

now that all the containers are working with new packages, we
have to cleanup the networking-bigswitch and get rid of the
packages that are redundant.

Change-Id: I7e3cdde12ac5809fcbbd3d0a219f9d6198c44eed