BCF-11522: Fix inconsitent unicode mode issue
- Initialize capabilities before using it
- Trigger capability update only once during server startup
  (Or at the first time OSP gets valid capabilities)
- Add more debug level messages

There are some issues with OSP threading, with periodic check, only
one thread would update capabilities despite there is only one
ServerPool instance. Currently we are not sure what causes this.

The workaround to this is to require customers to restart neutron
when they upgrade BCF, in which case, removing periodic capability
update would be more consistent.

Change-Id: I71b504e7b7b7f18efd08b9945161bef460390ec7